+++ frontend/src/Cabinet/utils/axiosUtils.ts
@@ -1,0 +1,8 @@
+import { InternalAxiosRequestConfig } from "axios";
+
+export const setAuthorizationHeader = (
+  config: InternalAxiosRequestConfig,
+  token?: string
+) => {
+  if (token) config.headers.set("Authorization", `Bearer ${token}`);
+};
